Reroute v3 urls in static build
