+++ modules/dashboard/Overview.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { useRouter } from 'next/router';
 import { useDelimitatedRoute } from '../../common/utils/router';
 import { SlowZonesWidget } from '../slowzones/SlowZonesWidget';
 import { SpeedWidget } from '../speed/SpeedWidget';
@@ -8,13 +7,12 @@ import { PageWrapper } from '../../common/layouts/PageWrapper';
 import { Layout } from '../../common/layouts/layoutTypes';
 import { RidershipWidget } from '../ridership/RidershipWidget';
 import { HEAVY_RAIL_LINES } from '../../common/types/lines';
+import { useRewriteV3Route } from '../../common/utils/middleware';
 
 export function Overview() {
   const { tab, line } = useDelimitatedRoute();
-  const router = useRouter();
 
-  // Semi-hacky way of making sure `/bus` works. Remove this when content is added to `/bus`.
-  if (line === 'line-bus') router.push('/bus/trips/single?busRoute=1');
+  useRewriteV3Route();
 
   const isHeavyRailLine = line ? HEAVY_RAIL_LINES.includes(line) : false;
   return (
